use friesian stringindex implementation for orca
