+++ python/friesian/src/bigdl/friesian/feature/table.py
@@ -34,6 +34,7 @@ from pyspark.sql.functions import col as pyspark_col, concat, udf, array, broadc
     lit, rank, monotonically_increasing_id, row_number, desc
 from pyspark.sql.types import ArrayType, DataType, StructType, StringType, StructField
 
+from bigdl.orca.data.transform import StringIndexer
 
 JAVA_INT_MIN = -2147483648
 JAVA_INT_MAX = 2147483647
@@ -889,43 +890,45 @@ class FeatureTable(Table):
         :return: A new FeatureTable which transforms categorical features into unique integer
                  values with provided StringIndexes.
         """
-        if not isinstance(columns, list):
-            columns = [columns]
-        if not isinstance(indices, list):
-            indices = [indices]
-        invalidInputError(len(columns) == len(indices),
-                          "columns len doesn't match indices lenngth")
-        if isinstance(indices[0], dict):
-            indices = list(map(lambda x: StringIndex.from_dict(x[1], columns[x[0]]),
-                               enumerate(indices)))
-        data_df = self.df
-        for i in range(len(columns)):
-            index_tbl = indices[i]
-            col_name = columns[i]
-            if broadcast:
-                index_tbl.broadcast()
-            if not do_split:
-                data_df = data_df.join(index_tbl.df, col_name, how="left") \
-                    .drop(col_name).withColumnRenamed("id", col_name)
-            else:
-                data_df = data_df.withColumn('row_id', F.monotonically_increasing_id())
-                tmp_df = data_df.select('row_id', col_name) \
-                    .withColumn(col_name, F.explode(F.split(F.col(col_name), sep)))
-                tmp_df = tmp_df.join(index_tbl.df, col_name, how="left") \
-                    .filter(F.col("id").isNotNull())
-                tmp_df = tmp_df.select('row_id', F.col("id"))
-                if keep_most_frequent:
-                    tmp_df = tmp_df.groupby('row_id') \
-                        .agg(F.array_sort(F.collect_list(F.col("id")))
-                             .getItem(0).alias("id"))
-                elif sort_for_array:
-                    tmp_df = tmp_df.groupby('row_id') \
-                        .agg(F.array_sort(F.collect_list(F.col("id"))).alias("id"))
-                else:
-                    tmp_df = tmp_df.groupby('row_id') \
-                        .agg(F.collect_list(F.col("id")).alias("id"))
-                data_df = data_df.join(tmp_df, 'row_id', 'left') \
-                    .drop('row_id').drop(col_name).withColumnRenamed("id", col_name)
+        data_df = StringIndexer.encode_string(self.df, columns, indices, broadcast, do_split,
+                      sep, sort_for_array, keep_most_frequent)
+        # if not isinstance(columns, list):
+        #     columns = [columns]
+        # if not isinstance(indices, list):
+        #     indices = [indices]
+        # invalidInputError(len(columns) == len(indices),
+        #                   "columns len doesn't match indices lenngth")
+        # if isinstance(indices[0], dict):
+        #     indices = list(map(lambda x: StringIndex.from_dict(x[1], columns[x[0]]),
+        #                        enumerate(indices)))
+        # data_df = self.df
+        # for i in range(len(columns)):
+        #     index_tbl = indices[i]
+        #     col_name = columns[i]
+        #     if broadcast:
+        #         index_tbl.broadcast()
+        #     if not do_split:
+        #         data_df = data_df.join(index_tbl.df, col_name, how="left") \
+        #             .drop(col_name).withColumnRenamed("id", col_name)
+        #     else:
+        #         data_df = data_df.withColumn('row_id', F.monotonically_increasing_id())
+        #         tmp_df = data_df.select('row_id', col_name) \
+        #             .withColumn(col_name, F.explode(F.split(F.col(col_name), sep)))
+        #         tmp_df = tmp_df.join(index_tbl.df, col_name, how="left") \
+        #             .filter(F.col("id").isNotNull())
+        #         tmp_df = tmp_df.select('row_id', F.col("id"))
+        #         if keep_most_frequent:
+        #             tmp_df = tmp_df.groupby('row_id') \
+        #                 .agg(F.array_sort(F.collect_list(F.col("id")))
+        #                      .getItem(0).alias("id"))
+        #         elif sort_for_array:
+        #             tmp_df = tmp_df.groupby('row_id') \
+        #                 .agg(F.array_sort(F.collect_list(F.col("id"))).alias("id"))
+        #         else:
+        #             tmp_df = tmp_df.groupby('row_id') \
+        #                 .agg(F.collect_list(F.col("id")).alias("id"))
+        #         data_df = data_df.join(tmp_df, 'row_id', 'left') \
+        #             .drop('row_id').drop(col_name).withColumnRenamed("id", col_name)
 
         return FeatureTable(data_df)
 
@@ -1000,36 +1003,39 @@ class FeatureTable(Table):
     def category_encode(self, columns, freq_limit=None, order_by_freq=False,
                         do_split=False, sep=',', sort_for_array=False, keep_most_frequent=False,
                         broadcast=True):
-        """
-        Category encode the given columns.
-
-        :param columns: str or a list of str, target columns to encode from string to index.
-        :param freq_limit: int, dict or None. Categories with a count/frequency below freq_limit
-               will be omitted from the encoding. Can be represented as either an integer,
-               dict. For instance, 15, {'col_4': 10, 'col_5': 2} etc. Default is None,
-               and in this case all the categories that appear will be encoded.
-        :param order_by_freq: boolean, whether the result StringIndex will assign smaller indices
-               to values with more frequencies. Default is False and in this case frequency order
-               may not be preserved when assigning indices.
-        :param do_split: bool, whether need to split column value to array to encode string.
-               Default is False.
-        :param sep: str, a string representing a regular expression to split a column value.
-               Default is ','.
-        :param sort_for_array: bool, whether need to sort array columns. Default is False.
-        :param keep_most_frequent: bool, whether need to keep most frequent value as the
-               column value. Default is False.
-        :param broadcast: bool, whether need to broadcast index when encode string.
-               Default is True.
-
-        :return: A tuple of a new FeatureTable which transforms categorical features into unique
-                 integer values, and a list of StringIndex for the mapping.
-        """
-        indices = self.gen_string_idx(columns, freq_limit=freq_limit, order_by_freq=order_by_freq,
-                                      do_split=do_split, sep=sep)
-        return self.encode_string(columns, indices, do_split=do_split, sep=sep,
-                                  sort_for_array=sort_for_array,
-                                  keep_most_frequent=keep_most_frequent,
-                                  broadcast=broadcast), indices
+        StringIndexer.category_encode(self.df, columns, freq_limit, order_by_freq,
+                        do_split, sep, sort_for_array, keep_most_frequent,
+                        broadcast)
+    #     """
+    #     Category encode the given columns.
+    #
+    #     :param columns: str or a list of str, target columns to encode from string to index.
+    #     :param freq_limit: int, dict or None. Categories with a count/frequency below freq_limit
+    #            will be omitted from the encoding. Can be represented as either an integer,
+    #            dict. For instance, 15, {'col_4': 10, 'col_5': 2} etc. Default is None,
+    #            and in this case all the categories that appear will be encoded.
+    #     :param order_by_freq: boolean, whether the result StringIndex will assign smaller indices
+    #            to values with more frequencies. Default is False and in this case frequency order
+    #            may not be preserved when assigning indices.
+    #     :param do_split: bool, whether need to split column value to array to encode string.
+    #            Default is False.
+    #     :param sep: str, a string representing a regular expression to split a column value.
+    #            Default is ','.
+    #     :param sort_for_array: bool, whether need to sort array columns. Default is False.
+    #     :param keep_most_frequent: bool, whether need to keep most frequent value as the
+    #            column value. Default is False.
+    #     :param broadcast: bool, whether need to broadcast index when encode string.
+    #            Default is True.
+    #
+    #     :return: A tuple of a new FeatureTable which transforms categorical features into unique
+    #              integer values, and a list of StringIndex for the mapping.
+    #     """
+    #     indices = self.gen_string_idx(columns, freq_limit=freq_limit, order_by_freq=order_by_freq,
+    #                                   do_split=do_split, sep=sep)
+    #     return self.encode_string(columns, indices, do_split=do_split, sep=sep,
+    #                               sort_for_array=sort_for_array,
+    #                               keep_most_frequent=keep_most_frequent,
+    #                               broadcast=broadcast), indices
 
     def one_hot_encode(self, columns, sizes=None, prefix=None, keep_original_columns=False):
         """
@@ -1126,107 +1132,109 @@ class FeatureTable(Table):
 
     def gen_string_idx(self, columns, freq_limit=None, order_by_freq=False,
                        do_split=False, sep=','):
-        """
-        Generate unique index value of categorical features. The resulting index would
-        start from 1 with 0 reserved for unknown features.
-
-        :param columns: str, dict or a list of str, dict, target column(s) to generate StringIndex.
-         dict is a mapping of source column names -> target column name if needs to combine multiple
-         source columns to generate index.
-         For example: {'src_cols':['a_user', 'b_user'], 'col_name':'user'}.
-        :param freq_limit: int, dict or None. Categories with a count/frequency below freq_limit
-               will be omitted from the encoding. Can be represented as either an integer,
-               dict. For instance, 15, {'col_4': 10, 'col_5': 2} etc. Default is None,
-               and in this case all the categories that appear will be encoded.
-        :param order_by_freq: boolean, whether the result StringIndex will assign smaller indices
-               to values with more frequencies. Default is False and in this case frequency order
-               may not be preserved when assigning indices.
-        :param do_split: bool, whether need to split column value to array to generate index.
-        Default is False.
-        :param sep: str, a string representing a regular expression to split a column value.
-         Default is ','.
-
-        :return: A StringIndex or a list of StringIndex.
-        """
-        if columns is None:
-            invalidInputError(False,
-                              "columns should be str or a list of str, but got None.")
-        is_single_column = False
-        if not isinstance(columns, list):
-            is_single_column = True
-            columns = [columns]
-        src_columns = []
-        for c in columns:
-            if isinstance(c, dict):
-                if 'src_cols' in c:
-                    src_columns.extend(c['src_cols'])
-            else:
-                src_columns.append(c)
-        check_col_exists(self.df, src_columns)
-        if freq_limit:
-            if isinstance(freq_limit, int):
-                freq_limit = str(freq_limit)
-            elif isinstance(freq_limit, dict):
-                freq_limit = ",".join(str(k) + ":" + str(v) for k, v in freq_limit.items())
-            else:
-                invalidInputError(False,
-                                  "freq_limit only supports int, dict or None, but get " +
-                                  freq_limit.__class__.__name__)
-        out_columns = []
-        simple_columns = []
-        df_id_list = []
-        for c in columns:
-            # union column
-            if isinstance(c, dict):
-                if 'src_cols' in c:
-                    src_cols = c['src_cols']
-                else:
-                    invalidInputError(False,
-                                      "Union columns must has argument 'src_cols'")
-                if 'col_name' in c:
-                    col_name = c['col_name']
-                else:
-                    col_name = src_cols[0] + '_union'
-                # process simple columns
-                if simple_columns:
-                    simple_df_id_list = generate_string_idx(self.df, simple_columns,
-                                                            freq_limit, order_by_freq)
-                    df_id_list.extend(simple_df_id_list)
-                    simple_columns = []
-                # process union columns
-                for i, src_c in enumerate(src_cols):
-                    if i == 0:
-                        dict_df = self.df.select(F.col(src_c).alias(col_name))
-                    else:
-                        dict_df = dict_df.union(self.df.select(F.col(src_c).alias(col_name)))
-                union_id_list = generate_string_idx(dict_df, [col_name],
-                                                    freq_limit, order_by_freq)
-                df_id_list.extend(union_id_list)
-                out_columns.append(col_name)
-            # single column
-            else:
-                if do_split:
-                    dict_df = self.df.select(F.col(c))
-                    dict_df = dict_df.withColumn(c, F.explode(F.split(c, sep)))
-                    split_id_list = generate_string_idx(dict_df, [c], freq_limit, order_by_freq)
-                    df_id_list.extend(split_id_list)
-                    out_columns.append(c)
-                else:
-                    simple_columns.append(c)
-                    out_columns.append(c)
-        if simple_columns:
-            simple_df_id_list = generate_string_idx(self.df, simple_columns,
-                                                    freq_limit, order_by_freq)
-            df_id_list.extend(simple_df_id_list)
-
-        string_idx_list = list(map(lambda x: StringIndex(x[0], x[1]),
-                                   zip(df_id_list, out_columns)))
-
-        # If input is a single column (not a list), then the output would be a single StringIndex.
-        if len(string_idx_list) == 1 and is_single_column:
-            return string_idx_list[0]
-        else:
-            return string_idx_list
+        StringIndexer.gen_string_idx(self.df, columns, freq_limit, order_by_freq,
+                       do_split, sep)
+    #     """
+    #     Generate unique index value of categorical features. The resulting index would
+    #     start from 1 with 0 reserved for unknown features.
+    #
+    #     :param columns: str, dict or a list of str, dict, target column(s) to generate StringIndex.
+    #      dict is a mapping of source column names -> target column name if needs to combine multiple
+    #      source columns to generate index.
+    #      For example: {'src_cols':['a_user', 'b_user'], 'col_name':'user'}.
+    #     :param freq_limit: int, dict or None. Categories with a count/frequency below freq_limit
+    #            will be omitted from the encoding. Can be represented as either an integer,
+    #            dict. For instance, 15, {'col_4': 10, 'col_5': 2} etc. Default is None,
+    #            and in this case all the categories that appear will be encoded.
+    #     :param order_by_freq: boolean, whether the result StringIndex will assign smaller indices
+    #            to values with more frequencies. Default is False and in this case frequency order
+    #            may not be preserved when assigning indices.
+    #     :param do_split: bool, whether need to split column value to array to generate index.
+    #     Default is False.
+    #     :param sep: str, a string representing a regular expression to split a column value.
+    #      Default is ','.
+    #
+    #     :return: A StringIndex or a list of StringIndex.
+    #     """
+    #     if columns is None:
+    #         invalidInputError(False,
+    #                           "columns should be str or a list of str, but got None.")
+    #     is_single_column = False
+    #     if not isinstance(columns, list):
+    #         is_single_column = True
+    #         columns = [columns]
+    #     src_columns = []
+    #     for c in columns:
+    #         if isinstance(c, dict):
+    #             if 'src_cols' in c:
+    #                 src_columns.extend(c['src_cols'])
+    #         else:
+    #             src_columns.append(c)
+    #     check_col_exists(self.df, src_columns)
+    #     if freq_limit:
+    #         if isinstance(freq_limit, int):
+    #             freq_limit = str(freq_limit)
+    #         elif isinstance(freq_limit, dict):
+    #             freq_limit = ",".join(str(k) + ":" + str(v) for k, v in freq_limit.items())
+    #         else:
+    #             invalidInputError(False,
+    #                               "freq_limit only supports int, dict or None, but get " +
+    #                               freq_limit.__class__.__name__)
+    #     out_columns = []
+    #     simple_columns = []
+    #     df_id_list = []
+    #     for c in columns:
+    #         # union column
+    #         if isinstance(c, dict):
+    #             if 'src_cols' in c:
+    #                 src_cols = c['src_cols']
+    #             else:
+    #                 invalidInputError(False,
+    #                                   "Union columns must has argument 'src_cols'")
+    #             if 'col_name' in c:
+    #                 col_name = c['col_name']
+    #             else:
+    #                 col_name = src_cols[0] + '_union'
+    #             # process simple columns
+    #             if simple_columns:
+    #                 simple_df_id_list = generate_string_idx(self.df, simple_columns,
+    #                                                         freq_limit, order_by_freq)
+    #                 df_id_list.extend(simple_df_id_list)
+    #                 simple_columns = []
+    #             # process union columns
+    #             for i, src_c in enumerate(src_cols):
+    #                 if i == 0:
+    #                     dict_df = self.df.select(F.col(src_c).alias(col_name))
+    #                 else:
+    #                     dict_df = dict_df.union(self.df.select(F.col(src_c).alias(col_name)))
+    #             union_id_list = generate_string_idx(dict_df, [col_name],
+    #                                                 freq_limit, order_by_freq)
+    #             df_id_list.extend(union_id_list)
+    #             out_columns.append(col_name)
+    #         # single column
+    #         else:
+    #             if do_split:
+    #                 dict_df = self.df.select(F.col(c))
+    #                 dict_df = dict_df.withColumn(c, F.explode(F.split(c, sep)))
+    #                 split_id_list = generate_string_idx(dict_df, [c], freq_limit, order_by_freq)
+    #                 df_id_list.extend(split_id_list)
+    #                 out_columns.append(c)
+    #             else:
+    #                 simple_columns.append(c)
+    #                 out_columns.append(c)
+    #     if simple_columns:
+    #         simple_df_id_list = generate_string_idx(self.df, simple_columns,
+    #                                                 freq_limit, order_by_freq)
+    #         df_id_list.extend(simple_df_id_list)
+    #
+    #     string_idx_list = list(map(lambda x: StringIndex(x[0], x[1]),
+    #                                zip(df_id_list, out_columns)))
+    #
+    #     # If input is a single column (not a list), then the output would be a single StringIndex.
+    #     if len(string_idx_list) == 1 and is_single_column:
+    #         return string_idx_list[0]
+    #     else:
+    #         return string_idx_list
 
     def _clone(self, df):
         return FeatureTable(df)
