+++ scala/friesian/src/main/scala/com/intel/analytics/bigdl/friesian/python/PythonFriesian.scala
@@ -91,62 +91,62 @@ class PythonFriesian[T: ClassTag](implicit ev: TensorNumeric[T]) extends PythonZ
     spark.createDataFrame(dfUpdated, schema)
   }
 
-  def generateStringIdx(df: DataFrame, columns: JList[String], frequencyLimit: String = null,
-                        orderByFrequency: Boolean = false)
-  : JList[DataFrame] = {
-    var default_limit: Option[Int] = None
-    val freq_map = scala.collection.mutable.Map[String, Int]()
-    if (frequencyLimit != null) {
-      val freq_list = frequencyLimit.split(",")
-      for (fl <- freq_list) {
-        val frequency_pair = fl.split(":")
-        if (frequency_pair.length == 1) {
-          default_limit = Some(frequency_pair(0).toInt)
-        } else if (frequency_pair.length == 2) {
-          freq_map += (frequency_pair(0) -> frequency_pair(1).toInt)
-        }
-      }
-    }
-    val cols = columns.asScala.toList
-    cols.map(col_n => {
-      val df_col = df
-        .select(col_n)
-        .filter(s"${col_n} is not null")
-        .groupBy(col_n)
-        .count()
-      val df_col_ordered = if (orderByFrequency) {
-        df_col.orderBy(col("count").desc)
-      } else df_col
-      val df_col_filtered = if (freq_map.contains(col_n)) {
-        df_col_ordered.filter(s"count >= ${freq_map(col_n)}")
-      } else if (default_limit.isDefined) {
-        df_col_ordered.filter(s"count >= ${default_limit.get}")
-      } else {
-        df_col_ordered
-      }
-
-      df_col_filtered.cache()
-      val count_list: Array[(Int, Int)] = df_col_filtered.rdd.mapPartitions(Utils.getPartitionSize)
-        .collect().sortBy(_._1)  // further guarantee prior partitions are given smaller indices.
-      val base_dict = scala.collection.mutable.Map[Int, Int]()
-      var running_sum = 0
-      for (count_tuple <- count_list) {
-        base_dict += (count_tuple._1 -> running_sum)
-        running_sum += count_tuple._2
-      }
-      val base_dict_bc = df_col_filtered.rdd.sparkContext.broadcast(base_dict)
-
-      val windowSpec = Window.partitionBy("part_id").orderBy(col("count").desc)
-      val df_with_part_id = df_col_filtered.withColumn("part_id", spark_partition_id())
-      val df_row_number = df_with_part_id.withColumn("row_number", row_number.over(windowSpec))
-      val get_label = udf((part_id: Int, row_number: Int) => {
-        row_number + base_dict_bc.value.getOrElse(part_id, 0)
-      })
-      df_row_number
-        .withColumn("id", get_label(col("part_id"), col("row_number")))
-        .drop("part_id", "row_number", "count")
-    }).asJava
-  }
+//  def generateStringIdx(df: DataFrame, columns: JList[String], frequencyLimit: String = null,
+//                        orderByFrequency: Boolean = false)
+//  : JList[DataFrame] = {
+//    var default_limit: Option[Int] = None
+//    val freq_map = scala.collection.mutable.Map[String, Int]()
+//    if (frequencyLimit != null) {
+//      val freq_list = frequencyLimit.split(",")
+//      for (fl <- freq_list) {
+//        val frequency_pair = fl.split(":")
+//        if (frequency_pair.length == 1) {
+//          default_limit = Some(frequency_pair(0).toInt)
+//        } else if (frequency_pair.length == 2) {
+//          freq_map += (frequency_pair(0) -> frequency_pair(1).toInt)
+//        }
+//      }
+//    }
+//    val cols = columns.asScala.toList
+//    cols.map(col_n => {
+//      val df_col = df
+//        .select(col_n)
+//        .filter(s"${col_n} is not null")
+//        .groupBy(col_n)
+//        .count()
+//      val df_col_ordered = if (orderByFrequency) {
+//        df_col.orderBy(col("count").desc)
+//      } else df_col
+//      val df_col_filtered = if (freq_map.contains(col_n)) {
+//        df_col_ordered.filter(s"count >= ${freq_map(col_n)}")
+//      } else if (default_limit.isDefined) {
+//        df_col_ordered.filter(s"count >= ${default_limit.get}")
+//      } else {
+//        df_col_ordered
+//      }
+//
+//      df_col_filtered.cache()
+//      val count_list: Array[(Int, Int)] = df_col_filtered.rdd.mapPartitions(Utils.getPartitionSize)
+//        .collect().sortBy(_._1)  // further guarantee prior partitions are given smaller indices.
+//      val base_dict = scala.collection.mutable.Map[Int, Int]()
+//      var running_sum = 0
+//      for (count_tuple <- count_list) {
+//        base_dict += (count_tuple._1 -> running_sum)
+//        running_sum += count_tuple._2
+//      }
+//      val base_dict_bc = df_col_filtered.rdd.sparkContext.broadcast(base_dict)
+//
+//      val windowSpec = Window.partitionBy("part_id").orderBy(col("count").desc)
+//      val df_with_part_id = df_col_filtered.withColumn("part_id", spark_partition_id())
+//      val df_row_number = df_with_part_id.withColumn("row_number", row_number.over(windowSpec))
+//      val get_label = udf((part_id: Int, row_number: Int) => {
+//        row_number + base_dict_bc.value.getOrElse(part_id, 0)
+//      })
+//      df_row_number
+//        .withColumn("id", get_label(col("part_id"), col("row_number")))
+//        .drop("part_id", "row_number", "count")
+//    }).asJava
+//  }
 
   def compute(df: DataFrame): Unit = {
     df.rdd.count()
